+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 4.11.3
 
+- feat: Added `ListTile.decorationMargin` parameter
 - fix: `TabView.newTabIcon` is now typed as a `Widget` ([#1187](https://github.com/bdlukaa/fluent_ui/issues/1187))
 - feat: `Expanded.enabled` ([#1188](https://github.com/bdlukaa/fluent_ui/issues/1188))
 - fix: Correctly apply `TextBox.unfocusedColor` on unfocused highlight border ([#1192](https://github.com/bdlukaa/fluent_ui/issues/1192))